pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.40",
    "pillow>=11.1.0",
    "ffmpeg-python>=0.2.0",
    "mysqlclient>=2.2.7",
    "python-socketio>=5.12.1",
    "aiortc>=1.11.0",
    "aiohttp>=3.11.14",
]
